Veja o PR #4 para notas de atualização.

+++ batservice.sh
@@ -1,6 +1,6 @@
-#!/system/bin/sh
+#!/bin/sh
 
-#    BatService v1.3 - battery conservation mode for Galaxy A20
+#    BatService v2.0 - battery conservation mode for Galaxy A20
 #
 #    Copyright (C) 2021 Cledson Ferreira
 #
@@ -18,100 +18,46 @@
 #    along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
 
-NAME="BATSERVICE"
-Name="BatService"
-name="batservice"
-VERSION="1.3.211112"
-
-
-if [ "$SERVICE_LIB" = "" ]; then
-  SERVICE_LIB="$PREFIX/lib/$name"
+if [ -z "$NO_PERMS" ]; then
+  . "$PREFIX/lib/batservice/env.rc"
 fi
 
-. "$SERVICE_LIB/perms.sh"
-if [ "$DATA" = "" ]; then
-  DATA="$PREFIX/etc/$name"
+NAME="BatService"
+VERSION="2.0.211205 beta"
+
+
+. "$LIB/perms.sh"
+. "$LIB/error.sh"
+. "$LIB/config.sh"
+. "$LIB/battery.sh"
+. "$LIB/jobs.sh"
+
+if [ ! -d "$DATA" ]; then
   backup_owner "$PREFIX"
   mkdir -p "$DATA"
   restore_owner "$PREFIX/etc"
   restore_owner "$DATA"
+  if [ -r "$EXIT_FILE" ]; then rm "$EXIT_FILE"; fi
 fi
 
-. "$SERVICE_LIB/error.sh"
-
-if [ "$NO_SERVICE" = "" ]; then
-  . "$SERVICE_LIB/startup-helper.sh"
-else
-  log_cleanup () { :; }
-fi
-
-. "$SERVICE_LIB/config.sh"
-. "$SERVICE_LIB/battery.sh"
-
-
-echo "$Name - conservação de bateria para o Galaxy A20"
+echo "$NAME - conservação de bateria para o Galaxy A20"
 echo "Versão $VERSION"
 echo
 
-if [ -r "$EXIT_FILE" ]; then
-  rm "$EXIT_FILE"
-fi
-
-prev_percent=0
-not_charging_set=$DISABLED
-
 echo " -*- STATUS DA BATERIA -*- "
 echo " $(date) "
-echo "  ============================="
+echo "=============================="
+
 
 while [ ! -r "$EXIT_FILE" ]; do
 
   config_refresh
 
-  battery_percent
-  battery_status
-  battery_current
-  battery_voltage
-  battery_temp
-
-  battery_switch_set get
-
-  if [ $prev_percent -ne $percent ]; then
-    battery_log
-    log_cleanup
-  fi
-
-  if ( [ "$status" = "Not charging" ] || [ "$status" = "Discharging" ] ); then
-
-    if ( [ $not_charging_set -eq $ENABLED ] && [ $percent -lt $MIN_PERCENT ] ); then
-      echo "ATIVAR carregamento"
-      battery_switch_set enable
-      not_charging_set=$DISABLED
-      echo
-      prev_percent=0
-      continue
-    fi
-
-    sleep $DELAY_REFRESH
-
-  elif [ "$status" = "Charging" ]; then
-
-    if ( [ $not_charging_set -eq $ENABLED ] || [ $percent -ge $MAX_PERCENT ] ); then
-      echo "DESATIVAR carregamento"
-      battery_switch_set disable
-      not_charging_set=$ENABLED
-      echo
-      prev_percent=0
-      continue
-    fi
-
-    sleep 6
-
-  fi
-
-  prev_percent=$percent
+  jobs_main
 
 done
+
+echo "=============================="
 
 battery_switch_set default
 rm "$EXIT_FILE"
